Move ABRP telemetry push from the app to the device driver
Each FordPass Connect app instance is already scoped to one vehicle
(singleInstance: false), so there was no real duplication risk to
avoid by keeping ABRP config on the app. The token is a per-vehicle
credential, so it belongs on the FordPass Vehicle device alongside
its other per-vehicle preferences (tire pressure/distance units)
rather than on the app.
